% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_plot}
\alias{add_plot}
\title{Add an interactive plot to a scene}
\usage{
add_plot(scene, type = NULL, encoding = NULL)
}
\arguments{
\item{scene}{A scene object}

\item{type}{A string representing the plot type}

\item{encoding}{Encoding of the variables}
}
\description{
Function that adds an interactive plot to a scene
}
