% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_scatterplot}
\alias{add_scatterplot}
\title{Add a scatterplot to an interactive scene}
\usage{
add_scatterplot(scene, encoding = NULL)
}
\arguments{
\item{scene}{A scene object}

\item{encoding}{Encoding of the variables: `v1` and `v2` (discrete or continuous),
`v3` continuous (optional)}
}
\description{
Function that adds a scatterplot to an interactive scene
}
