% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene.R
\name{set_scene}
\alias{set_scene}
\title{Create an Interactive Scene}
\usage{
set_scene(data, width = NULL, height = NULL, elementId = NULL, layout = NULL)
}
\arguments{
\item{data}{A dataframe (will be converted to JSON)}

\item{width}{Width of the scene}

\item{height}{Height of the scene}
}
\description{
Function that creates a skeleton of an interactive visual scene.
}
